requirements.txt
phidata
groq
streamlit
python-dotenv

